separate nav render logic to separate file
